Support Vector Machines
Selection of a model that predicts lawn mower ownership for the RisingMower.csv data. The study was done using SVM classification with linear, rbf, and polynomial kernels. The 'winning' model was saved to a pickle file (LinearSVM_lawnmower.pkl) which later was used to create a mini-application that asks the user for an income and lot size and answers if it's predicted that this property would own a lawnmower.